Extract helpers into own file
