+++ apps/web/src/components/types/types-chart/index.tsx
@@ -2,9 +2,9 @@
 
 import LoadingState from '@components/loading-state';
 import { client } from '@rq-client/index';
-import { capitalize } from '@utils/common';
 import { Type } from 'contract';
 import { Tooltip, TooltipContent, TooltipTrigger, TypeBadge, Typography } from 'ui';
+import { getDamageMultiplier, getTooltipText, sortEmptyTypeFirst } from './helpers';
 
 export function TypesChart(): JSX.Element {
   const { data, isFetching, isError, isLoading } = client.types.getAll.useQuery(['get-all-types']);
@@ -91,46 +91,4 @@ function Multiplier({ ofensiveType, defensiveType }: MultiplierProps): JSX.Eleme
       </TooltipContent>
     </Tooltip>
   );
-}
-
-// Aux
-function sortEmptyTypeFirst(typeA: Type, typeB: Type): number {
-  if (typeA.name === 'empty') {
-    return -1;
-  } else if (typeB.name === 'empty') {
-    return 1;
-  } else {
-    return 0;
-  }
-}
-
-type DamageMultiplierText = '0' | '2' | '1/2' | '';
-
-function getDamageMultiplier(ofensiveType: Type, defensiveType: Type): DamageMultiplierText {
-  if (ofensiveType.noDamageTo.includes(defensiveType.name)) {
-    return '0';
-  } else if (ofensiveType.doubleDamageTo.includes(defensiveType.name)) {
-    return '2';
-  } else if (ofensiveType.halfDamageTo.includes(defensiveType.name)) {
-    return '1/2';
-  }
-  return '';
-}
-
-function getTooltipText(ofensiveType: Type, defensiveType: Type): string {
-  const multiplier = getDamageMultiplier(ofensiveType, defensiveType);
-
-  const offensive = capitalize(ofensiveType.name);
-  const defensive = capitalize(defensiveType.name);
-
-  switch (multiplier) {
-    case '0':
-      return `${offensive} makes no damage to ${defensive}`;
-    case '2':
-      return `${offensive} is super effective against ${defensive}`;
-    case '1/2':
-      return `${offensive} is not very effective against ${defensive}`;
-    default:
-      return '';
-  }
 }
